add tab utility, more cleanups
